+++ sites/public/src/ListingView.tsx
@@ -1,3 +1,4 @@
+import React from "react"
 import ReactDOMServer from "react-dom/server"
 import Markdown from "markdown-to-jsx"
 import {
@@ -41,13 +42,12 @@ import {
   getSummariesTable,
   imageUrlFromListing,
   occupancyTable,
-  openDateState,
   t,
 } from "@bloom-housing/ui-components"
 import moment from "moment"
 import { ErrorPage } from "../pages/_error"
 import { useGetApplicationStatusProps } from "../lib/hooks"
-import { getGenericAddress } from "../lib/helpers"
+import { getGenericAddress, openInFuture } from "../lib/helpers"
 
 interface ListingProps {
   listing: Listing
@@ -62,6 +62,11 @@ export const ListingView = (props: ListingProps) => {
     content: appStatusContent,
     subContent: appStatusSubContent,
   } = useGetApplicationStatusProps(listing)
+
+  const appOpenInFuture = openInFuture(listing)
+  const hasNonReferralMethods = listing?.applicationMethods
+    ? listing.applicationMethods.some((method) => method.type !== ApplicationMethodType.Referral)
+    : false
 
   if (!listing) {
     return <ErrorPage />
@@ -274,7 +279,7 @@ export const ListingView = (props: ListingProps) => {
         applicationsDueDate={moment(listing.applicationDueDate).format(
           `MMM. DD, YYYY [${t("t.at")}] h A`
         )}
-        applicationsOpen={!openDateState(listing)}
+        applicationsOpen={!appOpenInFuture}
         applicationsOpenDate={moment(listing.applicationOpenDate).format("MMMM D, YYYY")}
         paperApplications={getPaperApplications()}
         paperMethod={!!getMethod(listing.applicationMethods, ApplicationMethodType.FileDownload)}
@@ -304,15 +309,6 @@ export const ListingView = (props: ListingProps) => {
   )
 
   const applicationsClosed = moment() > moment(listing.applicationDueDate)
-
-  //TODO: Add isReferralApplication boolean field to avoid this logic
-  const isReferralApp =
-    !listing.applicationDropOffAddress &&
-    !listing.applicationDropOffAddressType &&
-    !listing.applicationMailingAddress &&
-    !listing.applicationPickUpAddress &&
-    !listing.applicationPickUpAddressType &&
-    listing.applicationMethods?.length === 0
 
   return (
     <article className="flex flex-wrap relative max-w-5xl m-auto">
@@ -356,14 +352,14 @@ export const ListingView = (props: ListingProps) => {
             groupedUnits = byAMI ? getSummariesTable(byAMI.byUnitType) : []
 
             return (
-              <>
+              <React.Fragment key={percent}>
                 <h2 className="mt-4 mb-2">{t("listings.percentAMIUnit", { percent: percent })}</h2>
                 <GroupedTable
                   headers={unitSummariesHeaders}
                   data={groupedUnits}
                   responsiveCollapse={true}
                 />
-              </>
+              </React.Fragment>
             )
           })}
         {amiValues.length == 1 && (
@@ -381,7 +377,7 @@ export const ListingView = (props: ListingProps) => {
             event={lotteryResults}
             cloudName={process.env.cloudinaryCloudName}
           />
-          {!isReferralApp && !applicationsClosed ? <>{applySidebar()}</> : <></>}
+          {hasNonReferralMethods && !applicationsClosed ? <>{applySidebar()}</> : <></>}
         </div>
       </div>
       <ListingDetails>
@@ -486,12 +482,17 @@ export const ListingView = (props: ListingProps) => {
                 cloudName={process.env.cloudinaryCloudName}
               />
               {openHouseEvents && <OpenHouseEvent events={openHouseEvents} />}
-              {!isReferralApp && !applicationsClosed ? (
-                applySidebar()
-              ) : (
+              {hasNonReferralMethods && !applicationsClosed && applySidebar()}
+              {listing?.referralApplication && (
                 <ReferralApplication
-                  phoneNumber={t("application.referralApplication.phoneNumber")}
-                  description={t("application.referralApplication.instructions")}
+                  phoneNumber={
+                    listing.referralApplication.phoneNumber ||
+                    t("application.referralApplication.phoneNumber")
+                  }
+                  description={
+                    listing.referralApplication.externalReference ||
+                    t("application.referralApplication.instructions")
+                  }
                   title={t("application.referralApplication.furtherInformation")}
                 />
               )}
@@ -504,7 +505,7 @@ export const ListingView = (props: ListingProps) => {
             )}
             {lotterySection}
             <WhatToExpect listing={listing} />
-            <LeasingAgent listing={listing} />
+            {!appOpenInFuture && <LeasingAgent listing={listing} />}
           </aside>
         </ListingDetailItem>